==> notebooks/2_data_preparation.ipynb <==
import os
os.chdir('../')

## Imports

import pandas as pd
from src.data_preprocess import DataProcessor, aggregate_data

## Data Collect

df_imobs  = pd.read_csv("notebooks/outputs/df_imobs.csv").rename(columns={"dt_ativacao": "dt_calendar"})

df_imobs["dt_calendar"] = pd.to_datetime(df_imobs["dt_calendar"])
df_imobs = df_imobs.sort_values(["id_imobiliaria", "dt_calendar"])

print(df_imobs.shape)
df_imobs.head()

# Novas imobiliárias
(
    df_imobs["is_active"].apply(lambda x: (x < 100))
    .to_frame("new_imob")
    .groupby("new_imob").size()
    .to_frame("count")
)

df_imobs.groupby("id_imobiliaria", as_index=False)["is_active"].mean()

df_imobs["target"] = df_imobs["is_indemn_first_6months"].transform(
    lambda x: 1 if x >= 1 else 0 
)

## Shift das Features

df_imobs.columns

processor = DataProcessor(df_imobs)

# is_active já está com shift de 1 (já está com dado de um cohort atrás)
df_shift = processor.features_shift(
    group_by="id_imobiliaria",
    columns_to_shift={
        1: ["is_churn"],
        3: ["is_commun_first_90days"],
        4: ["is_debelado_first_4months"],
        6: ["exonerated_first_6months"]

    }
)

df_shift["churn_rate"] = df_shift["is_churn"] / df_shift["is_activated"]

def aggregate_data(df, column, window):
    return (
        (df.groupby(["id_imobiliaria"])[column]
        .rolling(window, min_periods=1).mean()
        .values) 
        # / 
        # (df.groupby(["id_imobiliaria"])["is_activated"]
        # .rolling(window, min_periods=1).mean()
        # .values)
    )

df_shift["agg_activated_last_90days"] = df_shift.groupby(["id_imobiliaria"])["is_activated"].rolling(6, min_periods=1).mean().values

aggregate_list = {
    # "is_activated": "agg_activated_last_90days",
    "is_commun_first_90days": "agg_comun_last_90days",
    # "is_debelado_first_4months": "agg_deb_last_4months",
    # "exonerated_first_6months": "agg_exon_last_6months",
    "churn_rate": "agg_churn_rate",
    # "rating_A": "agg_rating_A",
    # "rating_B": "agg_rating_B",
    # "rating_C": "agg_rating_C",
    # "rating_D": "agg_rating_D",
    # "rating_E": "agg_rating_E"
}

for to_agg, column in aggregate_list.items():
    df_shift[column] = (
        aggregate_data(df_shift, to_agg, 6)
    )

df_shift = df_shift.loc[df_shift["is_active"] > 100, :].reset_index(drop=True)

# 3,5% de target positivo
df_shift.groupby("target").size()

df_model = df_shift[~df_shift.exonerated_first_6months.isnull()]

df_model.dt_calendar.describe()

df_model.to_csv("notebooks/outputs/df_model_2.csv", index=False)

# df_new_imobs = (
#     df_shift.loc[df_shift["is_active"] <= 100, :][["dt_calendar", "id_imobiliaria"]]
#     .reset_index(drop=True)
#     .query("dt_calendar >= '2021-12' and dt_calendar <= '2022-05'")
#     .rename(columns={"dt_calendar": "dt_ativacao"})
# )

# df_new_imobs.to_csv("notebooks/outputs/df_new_imobs.csv", index=False)
