==> notebooks/3_optuna.ipynb <==
import os
os.chdir('../')

import json
import pandas as pd

from sklearn.metrics import roc_auc_score
from src.develop.hyperopt import OptunaOptimize

df_model = pd.read_csv("notebooks/outputs/df_model_2.csv")

df_model["dt_calendar"] = pd.to_datetime(df_model["dt_calendar"])
df_model = df_model[(df_model.dt_calendar >= "2020-01")]

print(df_model.shape)
df_model.head()


hyperopt = OptunaOptimize(
    df_model,
    from_="2021-06",
    to_="2021-11",
    eval_metric=roc_auc_score,
    trials=10
)
hyperopt.run()

print(f"BEST AUC: {hyperopt.best_result}")
hyperopt.best_params

path = "notebooks/outputs/params_et_2.json"

with open(path, "w") as f:
    json.dump(hyperopt.best_params, f)
