==> notebooks/6_econ_calculator.ipynb <==
import os
os.chdir('../')

## Imports

import pandas as pd
import datetime as dt

from risk_suite.economics import EconomicsCalculator

## Data Collect

#### Score Imobiliária

scores_4KST = (
    pd.read_csv("data/scores.csv")
    [["dt_ativacao", "id_contrato", "faixa_score"]]
    .rename(columns={
        "id_contrato": "contract_id",
        "faixa_score": "rating_4KST"
    })
)
scores_4KST.dt_ativacao = pd.to_datetime(scores_4KST["dt_ativacao"]).dt.strftime("%Y-%m") 

df_new_imobs = pd.read_csv("notebooks/outputs/df_new_imobs.csv")
df_bins = pd.read_csv("notebooks/outputs/df_bins_2.csv")

# Alinhando classificções com cohort de origem dos dados
# df_bins.dt_ativacao = (pd.to_datetime(df_bins["dt_ativacao"]) - pd.DateOffset(months=6)).dt.strftime("%Y-%m") 

df_new_imobs["optimal_segments"] = "4-novas"
df_bins = pd.concat([df_bins, df_new_imobs], ignore_index=True)
df_bins.dt_ativacao = (pd.to_datetime(df_bins["dt_ativacao"])).dt.strftime("%Y-%m") 

df_bins.head()

info_contracts = (
    pd.read_csv("data/info_contracts.csv")[["dt_ativacao", "id_imobiliaria", "id_contrato"]]
    .rename(columns={"id_contrato": "contract_id"})
)

info_contracts["dt_ativacao"] = pd.to_datetime(info_contracts["dt_ativacao"]).dt.strftime("%Y-%m")

df_bins.groupby("optimal_segments")["id_imobiliaria"].size()

df = (
    info_contracts
    .merge(df_bins, on=["dt_ativacao", "id_imobiliaria"])
    .merge(scores_4KST, on=["dt_ativacao", "contract_id"], how="left")
)
df.head()

#### Risco atual

calendar = pd.DataFrame(pd.date_range("2022-01", "2022-07", freq="M").strftime("%Y-%m"), columns=["dt_ativacao"])

old_score = pd.read_csv("data/old_risco_imob.csv")[["id_imobiliaria", "risco_imobiliaria"]].query("risco_imobiliaria != 0")

old_score["segments"] = (
    old_score["risco_imobiliaria"]
    .transform(
        lambda x: "A" if x == 2      else
                  "B" if x in (3, 4, 5, 6) else 
                  "C" if x == 7      else 
                  None
    )
)

old_score = old_score.merge(calendar, how="cross")
old_score.head()

old_score.groupby("segments")["id_imobiliaria"].size()

old_risk = info_contracts.merge(old_score[["dt_ativacao", "id_imobiliaria", "segments"]], on=["dt_ativacao", "id_imobiliaria"])
old_risk.head()

#### Calculadora

contracts = pd.read_parquet('../risk_suite/data/contracts.parquet')
defaults = pd.read_parquet('../risk_suite/data/defaults.parquet')
recoveries = pd.read_parquet('../risk_suite/data/recoveries.parquet')
revenues = pd.read_parquet('../risk_suite/data/revenues.parquet')

base_features = [
    'contract_id',
    'activation_date',
    'churn_date',
    'activation_month',
    'activation_quarter',
    'score_serasa',
    'rating',
    'rental_value',
]

contracts = contracts[base_features]
contracts = contracts.dropna(subset=['activation_date', 'rating'])
contracts = contracts.loc[lambda x: x['activation_quarter'] >= pd.Period('2020Q1')]
contracts = contracts.replace(['E1', 'E2', 'E3'], 'E')

## Auxiliary Functions

def report_economics2(aggkeys, revenues_pivot, defaults_pivot, recoveries_pivot):
    
    economics_df = pd.DataFrame(
        {
            'n_contracts': revenues_pivot.groupby(aggkeys).size(),
            'revenue_value': (revenues_pivot.sum(axis=1)).groupby(aggkeys).mean(),
            'prob_default': (defaults_pivot.sum(axis=1) > 0).groupby(aggkeys).mean(),
            'default_value': (defaults_pivot.sum(axis=1)).groupby(aggkeys).mean(),
            'recovery_value': (recoveries_pivot.sum(axis=1)).groupby(aggkeys).mean()
        }
    )

    economics_df = (
        economics_df
        .assign(recovery_efficiency=lambda x: x['recovery_value'] / x['default_value'])
        .assign(unit_economics=lambda x: x['revenue_value'] - x['default_value'] + x['recovery_value'])
        .assign(aggregate_margin=lambda x: x['unit_economics'] * x['n_contracts'])
    )

    return economics_df.sort_index(ascending=[True, False])

## Economics report

df = contracts.merge(df, on="contract_id")

econ_calculator = EconomicsCalculator(df, defaults, recoveries, revenues, max_history_date='2023-01')

defaults_pivot = econ_calculator._build_defaults_pivot()
recoveries_pivot = econ_calculator._build_recoveries_pivot()
revenues_pivot = econ_calculator._build_revenues_pivot()

#### Optimal Segmentation

report_optimal = report_economics2(["optimal_segments"], revenues_pivot, defaults_pivot, recoveries_pivot)
report_optimal

report_opt_rating = report_economics2(["optimal_segments", "rating_4KST"], revenues_pivot, defaults_pivot, recoveries_pivot)

pd.pivot_table(report_opt_rating, values="unit_economics", index="optimal_segments", columns="rating_4KST")

# Unit economics por segmento/contrato
pd.pivot_table(report_opt_rating, values="unit_economics", index="optimal_segments", columns="rating_4KST")

# Probabilidade de default por segmento/contrato
pd.pivot_table(report_opt_rating, values="prob_default", index="optimal_segments", columns="rating_4KST")

# Eficiência de Cobrança por segmento/contrato
pd.pivot_table(report_opt_rating, values="recovery_efficiency", index="optimal_segments", columns="rating_4KST")

# Mix de contratos por segmento/contrato
pd.pivot_table(report_opt_rating, values="n_contracts", index="optimal_segments", columns="rating_4KST")

#### Old Segmentation

report_old = report_economics2(["old_segments"], revenues_pivot, defaults_pivot, recoveries_pivot)
report_old

##### Transformed

report_old_transformed = report_economics2(["old_transformed_segments"], revenues_pivot, defaults_pivot, recoveries_pivot)
report_old_transformed

#### Risco Atual

old_risk_2 = contracts.merge(old_risk, on="contract_id")

econ_calculator_old = EconomicsCalculator(old_risk_2, defaults, recoveries, revenues, max_history_date='2023-01')

defaults_pivot_old = econ_calculator_old._build_defaults_pivot()
recoveries_pivot_old = econ_calculator_old._build_recoveries_pivot()
revenues_pivot_old = econ_calculator_old._build_revenues_pivot()

report_old_rating = report_economics2(["segments", "rating"], revenues_pivot_old, defaults_pivot_old, recoveries_pivot_old)

report_old_risk = report_economics2(["segments"], revenues_pivot_old, defaults_pivot_old, recoveries_pivot_old)
report_old_risk

# Probabilidade de default por segmento/contrato
pd.pivot_table(report_old_rating, values="n_contracts", index="segments", columns="rating")
